Simple Logger + New data storage + moved chunk code to it
All data should be collected in an object associated to one connection to a server, the "current" one.
The chunks/worlds are now stored in the new system, the old code references on that one now.
The spawn and re-spawn packet now trigger a new world to be saved.
A simple logger class has been added, which just does its job as a global formatter for output to console.