Getting rid of Error class, new exception printing
The Error class did for some reason not print any stacktrace not even a message. Further it was not able to catch the error, thus exchanged with a RuntimeException.
The Logger class now can print exceptions in a slightly different style.
Special handling for Auth server connection issues, to be enhanced!