BIG: Updating to good OpenGL mode, adding all received chunks to 3D
Using shaders now, TODO: Load them as resource instead of as file.
The 3D class has a queue for chunks to be loaded onto the GPU now. TODO: Better load order (shouldn't be as crappy as the original MC.
Added selfwritten-helper classes.
BlockModel now only delivers the color as array.
XYStorage can be cleared now, in case that the player switches the world.