Fix: The compression packet starts the play stage, not the encryption
Before the successful encryption did the switch to the play protocol. Causing the disconnect message, to not work if banned or not white listed. Causing the compression to fail if a packet with ID 3 comes in.
Now the encryption only enabled encryption. And the compression starts the play protocol. This is not fully correct, since some PlayerJoin packet actually does that job. But nah...