Adding possibility to modify existing chunks + make changes visible
Fix earlier commit. sub chunks are never null, but may be empty.
BlockUpdates can be done on sub chunks now.
The palette in the subchunks can be extended on demand. Even adding a bit. Reducing the bits is not (yet) possible. Always deserializes and serializes the whole sub chunk per edit.
3D will update dirty chunks now (in a primitive way).